FInished spec
I finished my spec. I talked to Mr. Broman in class about my spec today, and he told me that it would be a good idea if I wrote where the information was coming from for each page, in the spec. I did that in the remaining class period, and now I'm done with the spec.