Finished Settings Page
I finished the Settings page. It ended up requiring Data Binding, so it took a lot longer than I thought it would. However, I did end up adding both the sliders as well as the textblocks which updated as the sliders were updated. I am now on track for this week.