Changed structure of Play page
I changed the structure of the page. I wanted to fix the issue that the user could potentially not see that there exists a "Settings" page so I made it so that the play page was only accessible for the settings page. I spent the whole time thinking about this issue and implementing it. I also made it so that the minimum possible value of the settings sliders is 1, not 0, and also had to deal with the complications of fixing that.