Still trying to figure out how to use a GridView
I'm still trying to figure out how to use a GridView to have a list of buttons. In class, Mr. Broman told me that I could possibly use a ListView if I wanted. If I clicked one of them, I would pass it to the same method and there, I would decide which button was clicked and go from there. Unfortunately,  I did not make much real progress today, and I am not really advancing towards my goal for the end of the week.