Figured out how to create a GridView of Buttons
I created a GridView of Buttons successfully, and the number of buttons created is twice the number selected by the user, as desired. Right now, "1" is appended to the text of each button that is clicked in the Play tab to demonstrate that I can tell which button is clicked, but I have yet to find how to tell which Card each button belongs to.