Lots of debugging
I did a lot of debugging, and made the game feature complete. Now, there is a shufflebutton, so that once the user decides to stop looking at the values of the cards, they can reselect cards. I also fixed an instance where I used 0-based indices when I was supposed to use 1-based indices. Hopefully this is bugfree now but I don't really know.